fix: removed unused functions
